refactor(Header): modularize Header component into smaller components
Split the Header component into Logo, SearchBarWrapper, and UserActions components to improve maintainability and readability. This restructuring does not change the behavior but makes the codebase easier to manage.
